+++ example/render.js
@@ -1,0 +1,33 @@
+/**
+ * Test glyphs/borders/colors rendering
+ */
+const setup = require('./')
+
+
+setup({
+  positions:  [.5,.5, 1.5,.5, 2.5,.5, 3.5,.5, 4.5,.5, 5.5,.5, 6.5,.5, 7.5,.5, 8.5,.5, 9.5,.5,
+               .5,1.5, 1.5,1.5, 2.5,1.5,
+               .5,2.5, 1.5,2.5, 2.5,2.5,
+               .5,3.5, 1.5,3.5, 2.5,3.5, 3.5,3.5, 4.5,3.5, 5.5,3.5],
+  sizes:      [10, 20, 30, 40, 50, 60, 70, 80, 90, 100,
+               25, 30, 35,
+               40, 45, 50,
+               1, 2, 5, 10, 12, 15],
+  colors:     [0,0,0,1, .1,0,0,1, .2,0,0,1, .3,0,0,1, .4,0,0,1, .5,0,0,1, .6,0,0,1, .7,0,0,1, .8,0,0,1, .9,0,0,1,
+               1,0,0,1, 0,1,0,1, 0,0,1,1,
+               0,0,0,.2, 0,0,0,.5, 0,0,0,.8,
+               0,0,1,1, 0,0,1,1, 0,0,1,1, 0,0,1,1, 0,0,1,1, 0,0,1,1],
+  glyphs:     ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9',
+               '•', '+', '#',
+               'E', '=', 'mc²',
+               '●', '●', '●', '●', '●', '●'],
+  borderWidths: [1,1,1,1,1,1,1,1,1,1,
+                 2,2,2,
+                 0,0,0,
+                 .5, .5, .5, .5, .5, .5],
+  borderColors: [0,0,1,1, 0,0,1,.9, 0,0,1,.8, 0,0,1,.7, 0,0,1,.6, 0,0,1,.5, 0,0,1,.4, 0,0,1,.3, 0,0,1,.2, 0,0,1,.1,
+                 0,1,0,1, 0,0,1,1, 1,0,0,1,
+                 0,0,0,1, 0,0,0,1, 0,0,0,1,
+                 0,0,0,1, 0,0,0,1, 0,0,0,1, 0,0,0,1, 0,0,0,1, 0,0,0,1],
+})
+
